FreeBSD user functions make secure backups
The significant change is that in P:U:User:FreeBSD, the functions that back up the master.password file now back up to /var/backups (same as FreeBSD does) and set the permissions to 0. Added an option to disable automatic backups. -- suggested by Alan Ferrency

merged in latest Mail::Toaster::Utility, with an updated file_archive method and a few other API changes.